lib: organize src files
