+++ packages/lib-ts/src/types/ByteArray.ts
