+++ packages/lib-ts/src/types/Bytes.ts
