+++ packages/lib-ts/tests/types/ByteArray.spec.ts
@@ -1,4 +1,4 @@
-import { BigInt, ByteArray } from '../../common'
+import { BigInt, ByteArray } from '../../src/types'
 
 describe('ByteArray', () => {
   describe('empty', () => {
